fix:23 add refreshToken and add refreshEntity
