+++ src/main/resources/application-dev.yml
@@ -43,3 +43,4 @@ springdoc:
     disable-swagger-default-url: true
     display-query-params-without-oauth2: true
 
+
